Add build check for .env file

+++ scripts/build.sh
@@ -1,5 +1,13 @@
 #!/usr/bin/env bash
 
+ENV_FILE=.env
+
+## Make sure that the configuration file exists before continuing
+if [ ! -f "$ENV_FILE" ]; then
+  echo "$ENV_FILE does not exist"
+  exit 1
+fi
+
 next build
 next export
 echo 'roadmap.sh' > out/CNAME
